+ add all methods from SimpleCollisions to CollisionUtil - remove SimpleCollisions
